ai-integration/pyproject.toml
[project]
name = "ai-integration"
version = "0.1.0"
description = "AI Integration layer combining MCP Server and AI Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.1",
    "httpx>=0.28.1",
    "openai-agents>=0.6.4",
    "openai-agents-mcp>=0.0.8",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "mcp>=1.25.0",
]
